+++ vercel.json
@@ -9,7 +9,7 @@
             "use": "@vercel/next"
         },
         {
-            "src": "app/api/**/*.py",
+            "src": "app/api/main.py",
             "use": "@vercel/python",
             "config": {
                 "maxLambdaSize": "15mb",
@@ -19,12 +19,8 @@
     ],
     "routes": [
         {
-            "src": "/api/generate",
-            "dest": "app/api/generate.py"
-        },
-        {
-            "src": "/api/playlist",
-            "dest": "app/api/playlist.py"
+            "src": "/api/(.*)",
+            "dest": "app/api/main.py"
         },
         {
             "src": "/(.*)",
